requirements.txt
boto3~=1.10.1
click~=7.0
parallelpipe~=0.2.6
pre-commit~=1.18.3
psycopg2~=2.8.4
pyproj~=2.4.0
pytest~=5.2.1
pyyaml~=5.1.2
rasterio[s3]~=1.1.0
shapely~=1.6.4.post2
